README.md: Adjust headings
I originally made these effectively <h2> headers because I didn't
want to compete with the GitHub headings when the file was viewed
on their web interface. As it turns out, GitHub renders this in an
HTML5 article tag, which sits in its own world on the page, so it
shouldn't matter.

Signed-off-by: Alan Orth <alan.orth@gmail.com>